fix: handle the case when array is passed in getValidJson method
